* updated to v0_3_7 * Changed to glog

+++ source/opentm2/core/utilities/FilesystemHelper.cpp
@@ -76,8 +76,10 @@ FILE* FilesystemHelper::CreateFile(const std::string& path, const std::string& m
         || (strcasestr(cpath, ".TMD") )
         //|| (strcasestr(cpath, ".MEM") )
     ){
-        LogMessage4(INFO, "filesystem_open_file::Openning data file(with ext. .TMI, .TMD, .MEM => forcing to use filebuffers, fName = ", 
+        if(VLOG_IS_ON(1)){
+            LogMessage4(INFO, "filesystem_open_file::Openning data file(with ext. .TMI, .TMD, .MEM => forcing to use filebuffers, fName = ", 
                         cpath, ", useFilebuffer before = ", toStr(useBuffer).c_str());
+        }
         useBuffer = true;
     }
     return OpenFile(path, mode, useBuffer);
@@ -92,7 +94,9 @@ int FilesystemHelper::MoveFile(std::string oldPath, std::string newPath){
        LogMessage6(ERROR, "MoveFile:: cannot move ", fixedOldPath.c_str()," to ", fixedNewPath.c_str(), ", error = ", strerror(errno));
        return errno;
     }else{
-        LogMessage4(INFO,"MoveFile:: file moved from ", fixedOldPath.c_str()," to ", fixedNewPath.c_str());
+        if(VLOG_IS_ON(1)){
+            LogMessage4(INFO,"MoveFile:: file moved from ", fixedOldPath.c_str()," to ", fixedNewPath.c_str());
+        }
     }
     return 0;
 }
@@ -102,7 +106,9 @@ FILE* FilesystemHelper::OpenFile(const std::string& path, const std::string& mod
     fixedPath = FixPath(fixedPath);
     long fileSize = 0;
     FileBuffer* pFb = NULL;
-    LogMessage6(INFO, "FilesystemHelper::OpenFile, path =", path.c_str(), ", mode = ", mode.c_str(), ", useBuffer = ", toStr(useBuffer).c_str());
+    if(VLOG_IS_ON(1)){
+        LogMessage6(DEBUG, "FilesystemHelper::OpenFile, path =", path.c_str(), ", mode = ", mode.c_str(), ", useBuffer = ", toStr(useBuffer).c_str());
+    }
     FILE *ptr = fopen(fixedPath.c_str(), mode.c_str());
     if(!ptr){
         LogMessage2(ERROR, "FilesystemHelper::OpenFile:: can't open file ", path.c_str());
@@ -112,31 +118,43 @@ FILE* FilesystemHelper::OpenFile(const std::string& path, const std::string& mod
         }
 
         if(getFileBufferInstance()->find(fixedPath) != getFileBufferInstance()->end()){
-            LogMessage2(WARNING, "OpenFile::Filebuffer wasn't created, it's already exists for ", fixedPath.c_str());
+            //if(VLOG_IS_ON(1)){
+                LogMessage2(WARNING, "OpenFile::Filebuffer wasn't created, it's already exists for ", fixedPath.c_str());
+            //}
             pFb = &(*getFileBufferInstance())[fixedPath];
             pFb->offset = 0;
         }else{
             pFb = &(*getFileBufferInstance())[fixedPath];
             if( fileSize > 0 ){
                 pFb->data.resize(fileSize);
-                LogMessage4(INFO, "OpenFile:: file size >0  -> Filebuffer resized to filesize(",toStr(fileSize).c_str(),"), fname = ", fixedPath.c_str());
+                
+                if(VLOG_IS_ON(1)){
+                    LogMessage4(INFO, "OpenFile:: file size >0  -> Filebuffer resized to filesize(",toStr(fileSize).c_str(),"), fname = ", fixedPath.c_str());
+                }
                 int readed = fread(&pFb->data[0], fileSize, 1, ptr);
-                LogMessage6(INFO, "OpenFile:: file size >0  -> Filebuffer reading to buffer, size = ",toStr(fileSize).c_str(),
+                if(VLOG_IS_ON(1)){
+                    LogMessage6(INFO, "OpenFile:: file size >0  -> Filebuffer reading to buffer, size = ",toStr(fileSize).c_str(),
                             "), fname = ", fixedPath.c_str(), "; readed = ", toStr(readed).c_str());
+                }
             }else{
-                LogMessage2(INFO, "OpenFile:: file size <=0  -> Filebuffer resized to default value, fname = ", fixedPath.c_str());
+                if(VLOG_IS_ON(1)){
+                    LogMessage2(INFO, "OpenFile:: file size <=0  -> Filebuffer resized to default value, fname = ", fixedPath.c_str());
+                }
                 pFb->data.resize(16384);
             }
-            LogMessage4(INFO, "OpenFile::Filebuffer created for file ", fixedPath.c_str(), " with size = ", toStr(pFb->data.size()).c_str());
+            if(VLOG_IS_ON(1)){
+                LogMessage4(INFO, "OpenFile::Filebuffer created for file ", fixedPath.c_str(), " with size = ", toStr(pFb->data.size()).c_str());
+            }
             //pFb->status = FileBufferStatus::OPEN;
         }
 
         pFb->offset = 0;
     }
 
-    LogMessage8(DEBUG, "FilesystemHelper::OpenFile():: path = ", fixedPath.c_str(), "; mode = ", mode.c_str(), "; ptr = ", toStr((long int)ptr).c_str(),
+    if(VLOG_IS_ON(1)){
+        LogMessage8(DEBUG, "FilesystemHelper::OpenFile():: path = ", fixedPath.c_str(), "; mode = ", mode.c_str(), "; ptr = ", toStr((long int)ptr).c_str(),
              "size = ", toStr(fileSize).c_str());
-
+    }
     if(ptr == NULL){
         __last_error_code = FILEHELPER_FILE_PTR_IS_NULL;
     }
@@ -169,7 +187,9 @@ int FilesystemHelper::DeleteFile(const std::string& path){
         LogMessage4(ERROR, "FilesystemHelper::DeleteFile(",fixedPath.c_str() , ") ERROR res = ", toStr(errCode).c_str());
         return errCode;
     }else{
-        LogMessage3(DEBUG, "FilesystemHelper::DeleteFile(",fixedPath.c_str() , ") res = FILEHELPER_NO_ERROR");
+        if(VLOG_IS_ON(1)){
+            LogMessage3(DEBUG, "FilesystemHelper::DeleteFile(",fixedPath.c_str() , ") res = FILEHELPER_NO_ERROR");
+        }
         return FILEHELPER_NO_ERROR;
     }
 }
@@ -180,7 +200,9 @@ bool FilesystemHelper::FileExists(std::string&& path){
     fixedPath = FixPath(fixedPath);
     struct stat buffer;   
     bool exists = (stat (fixedPath.c_str(), &buffer) == 0); 
-    LogMessage5(INFO, __func__, ":: file \'", fixedPath.c_str(),"\' returned ", toStr(exists));
+    if(VLOG_IS_ON(1)){
+        LogMessage5(INFO, __func__, ":: file \'", fixedPath.c_str(),"\' returned ", toStr(exists));
+    }
     return exists;
 }
 
@@ -194,12 +216,14 @@ int FilesystemHelper::RemoveDirWithFiles(const std::string& path){
     system(command.c_str());
     
     if(int errCode = system(command.c_str()) ){
-        if(CheckLogLevel(DEBUG)){
+        if(CheckLogLevel(DEBUG) && VLOG_IS_ON(1)){
             LogMessage4(ERROR, "FilesystemHelper::RemoveDirWithFiles(",fixedPath.c_str() , ") ERROR res = ", toStr(errCode).c_str());
         }
         return errCode;
     }else{
-        LogMessage3(DEBUG, "FilesystemHelper::RemoveDirWithFiles(",fixedPath.c_str() , ") res = FILEHELPER_NO_ERROR");
+        if(VLOG_IS_ON(1)){
+            LogMessage3(DEBUG, "FilesystemHelper::RemoveDirWithFiles(",fixedPath.c_str() , ") res = FILEHELPER_NO_ERROR");
+        }
         return FILEHELPER_NO_ERROR;
     }
 }
@@ -224,8 +248,10 @@ int FilesystemHelper::WriteToBuffer(FILE *& ptr, const void* buff, const int buf
 
         pFb->status |= MODIFIED;
         if(offset + buffSize > pFb->data.size()){
-            LogMessage6(DEBUG, "FilesystemHelper::WriteToBuffer::Resizing file ", fName.c_str()," from ", toStr(pFb->data.size()).c_str(),
+            if(VLOG_IS_ON(1)){
+                LogMessage6(DEBUG, "FilesystemHelper::WriteToBuffer::Resizing file ", fName.c_str()," from ", toStr(pFb->data.size()).c_str(),
                     " to ", toStr(offset+buffSize).c_str());
+            }
             pFb->data.resize(offset + buffSize);
         }
 
@@ -236,7 +262,9 @@ int FilesystemHelper::WriteToBuffer(FILE *& ptr, const void* buff, const int buf
         LogMessage2(ERROR, "FilesystemHelper::WriteToBuffer:: can't find buffer for file ", fName.c_str());
         return -1;
     }
-    LogMessage4(DEBUG, "FilesystemHelper::WriteToBuffer:: success, ", toStr(buffSize).c_str()," bytes written to buffer for ", fName.c_str());
+    if(VLOG_IS_ON(1)){
+        LogMessage4(DEBUG, "FilesystemHelper::WriteToBuffer:: success, ", toStr(buffSize).c_str()," bytes written to buffer for ", fName.c_str());
+    }
     return 0;
 }
 
@@ -246,7 +274,9 @@ int FilesystemHelper::ReadBuffer(FILE*& ptr, void* buff, const int buffSize, int
     FileBuffer* pFb = NULL;
 
     if(getFileBufferInstance()->find(fName) == getFileBufferInstance()->end()){
-        LogMessage2(DEVELOP,"ReadBuffer:: file not found in buffers, fName = ", fName.c_str());
+        if(VLOG_IS_ON(1)){
+            LogMessage2(DEVELOP,"ReadBuffer:: file not found in buffers, fName = ", fName.c_str());
+        }
         return FILEHELPER_WARNING_BUFFER_FOR_FILE_NOT_OPENED;
     }
     pFb = &(*getFileBufferInstance())[fName];
@@ -262,14 +292,17 @@ int FilesystemHelper::ReadBuffer(FILE*& ptr, void* buff, const int buffSize, int
     }
     PUCHAR p = &(pFb->data[offset]);
     //
-    if(CheckLogLevel(DEVELOP)){
+    if(VLOG_IS_ON(1) && CheckLogLevel(DEVELOP)){
         std::string msg = std::string(__func__) + ":: fName = " + fName + "; buff size = " + toStr(buffSize) + "; data.size = " + toStr(pFb->data.size()) + "; offset = "+ toStr(offset) + ";";
         LogMessage(DEVELOP, msg.c_str());
+        //LOG_DEVELOP_MSG << msg;
     }
     //
     memcpy(buff, p, buffSize);
     bytesRead = buffSize;
-    LogMessage6(DEBUG, "ReadBuffer::", toStr(buffSize).c_str()," bytes read from buffer to ", fName.c_str(), " starting from ", toStr(offset).c_str() );
+    if(VLOG_IS_ON(1)){
+        LogMessage6(DEBUG, "ReadBuffer::", toStr(buffSize).c_str()," bytes read from buffer to ", fName.c_str(), " starting from ", toStr(offset).c_str() );
+    }
     return 0;
 }
 
@@ -283,7 +316,9 @@ int FilesystemHelper::FlushBufferIntoFile(const std::string& fName){
         (*pFBs)[fName].status &= ~MODIFIED;// reset modified flag
         //pFBs->erase(fName);
     }else{
-        LogMessage2(INFO,"FilesystemHelper::FlushBufferIntoFile:: filebuffer not found, fName = ", fName.c_str());
+        if(VLOG_IS_ON(1)){
+            LogMessage2(INFO,"FilesystemHelper::FlushBufferIntoFile:: filebuffer not found, fName = ", fName.c_str());
+        }
     }
     return 0;
 }
@@ -295,7 +330,9 @@ int FilesystemHelper::WriteBuffToFile(std::string fName){
         pFb = &(*pFBs)[fName];
 
         if(pFb->status & FileBufferStatus::MODIFIED){
-            LogMessage(INFO, "WriteBuffToFile:: writing files from buffer");
+            if(VLOG_IS_ON(1)){
+                LogMessage(INFO, "WriteBuffToFile:: writing files from buffer");
+            }
             PUCHAR bufStart = &pFb->data[0];
             int size = pFb->data.size();
             
@@ -303,7 +340,9 @@ int FilesystemHelper::WriteBuffToFile(std::string fName){
             WriteToFile(ptr, bufStart, size);
             fclose(ptr);
         }else{
-            LogMessage2(INFO,"WriteBuffToFile:: buffer not modified, so no need to overwrite file, fName = ", fName.c_str());
+            if(VLOG_IS_ON(1)){
+                LogMessage2(INFO,"WriteBuffToFile:: buffer not modified, so no need to overwrite file, fName = ", fName.c_str());
+            }
         }
     }else{
         LogMessage2(ERROR,"WriteBuffToFile:: buffer not found, fName = ", fName.c_str());
@@ -347,10 +386,14 @@ int FilesystemHelper::FlushAllBuffers(std::string * modifiedFiles){
 }
 
 int FilesystemHelper::CloseFile(FILE*& ptr){
-    LogMessage(DEBUG, "called FilesystemHelper::CloseFile()");
+    if(VLOG_IS_ON(1)){
+        LogMessage(DEBUG, "called FilesystemHelper::CloseFile()");
+    }
     if(ptr){
         std::string fName = GetFileName(ptr);
-        LogMessage4(DEBUG, "FilesystemHelper::CloseFile(", toStr((long int) ptr).c_str() , "), fName = ", fName.c_str() );
+        if(VLOG_IS_ON(1)){
+            LogMessage4(DEBUG, "FilesystemHelper::CloseFile(", toStr((long int) ptr).c_str() , "), fName = ", fName.c_str() );
+        }
         fclose(ptr);
         FlushBufferIntoFile(fName);
         CloseFileBuffer(fName);
@@ -426,9 +469,10 @@ std::vector<std::string> FilesystemHelper::GetFilesList(std::string&& directory)
         LogMessage3(ERROR,__func__,":: can't open dir, path = ", directory.c_str());
         return {};
     }
-
-    LogMessage5(INFO,__func__,":: returned ",toStr(files.size())," files,  path = ", directory.c_str());
-    return files;
+    if(VLOG_IS_ON(1)){    
+        LogMessage5(INFO,__func__,":: returned ",toStr(files.size())," files,  path = ", directory.c_str());
+    }
+return files;
 }
 
 std::vector<std::string> FilesystemHelper::FindFiles(const std::string& name){
@@ -463,8 +507,7 @@ std::vector<std::string> FilesystemHelper::FindFiles(const std::string& name){
             }
         }
         closedir (dir);
-    } else {
-
+    } else {      
         LogMessage3(ERROR, "FilesystemHelper::FindFiles:: dir = ",dirPath.c_str() , "; can't open directory");
         __last_error_code = FILEHELPER_ERROR_CANT_OPEN_DIR;
     }
@@ -504,16 +547,19 @@ int FilesystemHelper::WriteToFileBuff(FILE*& ptr, const void* buff, const long u
 
     if(getFileBufferInstance()->find(fName) != getFileBufferInstance()->end()){
 
-        LogMessage6(DEBUG,"Writing ", toStr(buffSize).c_str(), " bytes to filebuffer ", fName.c_str(),
+        if(VLOG_IS_ON(1)){
+            LogMessage6(DEBUG,"Writing ", toStr(buffSize).c_str(), " bytes to filebuffer ", fName.c_str(),
             " starting from position ", toStr(startingPosition).c_str());
+        }
         WriteToBuffer(ptr, buff, buffSize, startingPosition);
         iBytesWritten = buffSize;
     }else{
-        LogMessage4(DEBUG, "File is not opened in filebuffers-> writting to file, fName = ", fName.c_str(), ", fId = ", toStr((long)ptr).c_str());
+        if(VLOG_IS_ON(1)){
+            LogMessage4(DEBUG, "File is not opened in filebuffers-> writting to file, fName = ", fName.c_str(), ", fId = ", toStr((long)ptr).c_str());
 
-        LogMessage6(DEBUG,"Writing ", toStr(buffSize).c_str(), " bytes to file ", fName.c_str(),
+            LogMessage6(DEBUG,"Writing ", toStr(buffSize).c_str(), " bytes to file ", fName.c_str(),
             " starting from position ", toStr(startingPosition).c_str());
-
+        }
         #ifndef TEMPORARY_HARDCODED
         if(CheckLogLevel(DEBUG)){
             std::string filesInBuf = "FilesystemHelper::WriteToFileBuff:: files in filebuffer:\n";
@@ -521,7 +567,9 @@ int FilesystemHelper::WriteToFileBuff(FILE*& ptr, const void* buff, const long u
             {
                 filesInBuf += it->first + '\n';
             }
-            LogMessage(DEBUG, filesInBuf.c_str());
+            if(VLOG_IS_ON(1)){
+                LogMessage(DEBUG, filesInBuf.c_str());
+            }
         }
 
         #endif 
@@ -799,7 +847,7 @@ int FilesystemHelper::ReadFileBuff(FILE*& ptr, void* buff, const int buffSize, i
         }
 
     }else{
-        LogMessage2(INFO, "FilesystemHelper::ReadFileBuff, file buff not found-> using reading directly from file, fName = ", fName.c_str());
+        LogMessage2(DEBUG, "FilesystemHelper::ReadFileBuff, file buff not found-> using reading directly from file, fName = ", fName.c_str());
         errCode = ReadFile(ptr, buff, buffSize, bytesRead);
     }
     return errCode;
@@ -813,7 +861,7 @@ long FnGetFileSize(std::string&& filename)
 {
     struct stat stat_buf;
     int rc = stat(filename.c_str(), &stat_buf);
-    int logLevel = rc ==0? INFO : ERROR;
+    int logLevel = rc ==0? DEBUG : ERROR;
     LogMessage6(logLevel, "FnGetFileSize:: for ", filename.c_str(), " is ", toStr(stat_buf.st_size).c_str()," , rc = ", toStr(rc).c_str() );
     return rc == 0 ? stat_buf.st_size : -1;
 }
